samsariya-bot/requirements.txt
### requirements.txt

python-telegram-bot==20.3
fastapi==0.95.2
uvicorn==0.22.0
pydantic<2.0.0,>=1.6.2
python-dotenv==1.0.0